Attempt to turn on additional build tests
