+++ cmd/podman/containers/create.go
@@ -62,6 +62,9 @@ func createFlags(flags *pflag.FlagSet) {
 		_ = flags.MarkHidden("env-host")
 		_ = flags.MarkHidden("http-proxy")
 	}
+	// Not sure we want these exposed yet.  If we do, they need to be documented in man pages
+	_ = flags.MarkHidden("override-arch")
+	_ = flags.MarkHidden("override-os")
 }
 
 func init() {
