+++ cmd/podman/system/service.go
@@ -26,13 +26,12 @@ Enable a listening service for API access to Podman commands.
 `
 
 	srvCmd = &cobra.Command{
-		Use:   "service [flags] [URI]",
-		Args:  cobra.MaximumNArgs(1),
-		Short: "Run API service",
-		Long:  srvDescription,
-		RunE:  service,
-		Example: `podman system service --time=0 unix:///tmp/podman.sock
-  podman system service --varlink --time=0 unix:///tmp/podman.sock`,
+		Use:     "service [flags] [URI]",
+		Args:    cobra.MaximumNArgs(1),
+		Short:   "Run API service",
+		Long:    srvDescription,
+		RunE:    service,
+		Example: `podman system service --time=0 unix:///tmp/podman.sock`,
 	}
 
 	srvArgs = struct {
@@ -50,7 +49,6 @@ func init() {
 
 	flags := srvCmd.Flags()
 	flags.Int64VarP(&srvArgs.Timeout, "time", "t", 5, "Time until the service session expires in seconds.  Use 0 to disable the timeout")
-	flags.Int64Var(&srvArgs.Timeout, "timeout", 5, "Time until the service session expires in seconds.  Use 0 to disable the timeout")
 	flags.BoolVar(&srvArgs.Varlink, "varlink", false, "Use legacy varlink service instead of REST")
 
 	_ = flags.MarkDeprecated("varlink", "valink API is deprecated.")
