--- docs/source/markdown/links/podman-image-diff.1
@@ -1,1 +1,0 @@
-.so man1/podman-diff.1
