+++ docs/source/markdown/podman-image.1.md
@@ -14,6 +14,7 @@ The image command allows you to manage images
 | Command  | Man Page                                        | Description                                                                 |
 | -------- | ----------------------------------------------- | --------------------------------------------------------------------------- |
 | build    | [podman-build(1)](podman-build.1.md)            | Build a container using a Dockerfile.                                       |
+| diff     | [podman-image-diff(1)](podman-image-diff.1.md)  | Inspect changes on an image's filesystem.                                   |
 | exists   | [podman-image-exists(1)](podman-image-exists.1.md) | Check if an image exists in local storage.                               |
 | history  | [podman-history(1)](podman-history.1.md)        | Show the history of an image.                                               |
 | import   | [podman-import(1)](podman-import.1.md)          | Import a tarball and save it as a filesystem image.                         |
@@ -25,6 +26,7 @@ The image command allows you to manage images
 | push     | [podman-push(1)](podman-push.1.md)              | Push an image from local storage to elsewhere.                              |
 | rm       | [podman-rmi(1)](podman-rmi.1.md)                | Removes one or more locally stored images.                                  |
 | save     | [podman-save(1)](podman-save.1.md)              | Save an image to docker-archive or oci.                                     |
+| search   | [podman-search(1)](podman-search.1.md)          | Search a registry for an image.                                             |
 | sign     | [podman-image-sign(1)](podman-image-sign.1.md)  | Create a signature for an image.                                            |
 | tag      | [podman-tag(1)](podman-tag.1.md)                | Add an additional name to a local image.                                    |
 | untag    | [podman-untag(1)](podman-untag.1.md)            | Removes one or more names from a locally-stored image.                      |
