+++ docs/source/markdown/podman-manifest-push.1.md
@@ -19,7 +19,7 @@ The list image's ID and the digest of the image's manifest.
 Push the images mentioned in the manifest list or image index, in addition to
 the list or index itself.
 
-**--authfile** *path*
+**--authfile**=*path*
 
 Path of the authentication file. Default is ${XDG\_RUNTIME\_DIR}/containers/auth.json, which is set using `podman login`.
 If the authorization state is not found there, $HOME/.docker/config.json is checked, which is set using `docker login`. (Not available for remote commands)
@@ -27,22 +27,22 @@ If the authorization state is not found there, $HOME/.docker/config.json is chec
 Note: You can also override the default path of the authentication file by setting the REGISTRY\_AUTH\_FILE
 environment variable. `export REGISTRY_AUTH_FILE=path`
 
-**--cert-dir** *path*
+**--cert-dir**=*path*
 
 Use certificates at *path* (\*.crt, \*.cert, \*.key) to connect to the registry.
 Default certificates directory is _/etc/containers/certs.d_. (Not available for remote commands)
 
-**--creds** *creds*
+**--creds**=*creds*
 
 The [username[:password]] to use to authenticate with the registry if required.
 If one or both values are not supplied, a command line prompt will appear and the
 value can be entered.  The password is entered without echo.
 
-**--digestfile** *Digestfile*
+**--digestfile**=*Digestfile*
 
 After copying the image, write the digest of the resulting image to the file.
 
-**--format, -f**
+**--format**, **-f**=*format*
 
 Manifest list type (oci or v2s2) to use when pushing the list (default is oci).
 
@@ -50,15 +50,19 @@ Manifest list type (oci or v2s2) to use when pushing the list (default is oci).
 
 Delete the manifest list or image index from local storage if pushing succeeds.
 
+**--quiet**, **-q**
+
+When writing the manifest, suppress progress output
+
 **--remove-signatures**
 
 Don't copy signatures when pushing images.
 
-**--sign-by** *fingerprint*
+**--sign-by**=*fingerprint*
 
 Sign the pushed images using the GPG key that matches the specified fingerprint.
 
-**--tls-verify** *bool-value*
+**--tls-verify**
 
 Require HTTPS and verify certificates when talking to container registries (defaults to true) (Not available for remote commands)
