+++ docs/source/markdown/podman-network-inspect.1.md
@@ -10,10 +10,6 @@ podman\-network\-inspect - Displays the raw CNI network configuration for one or
 Display the raw (JSON format) network configuration. This command is not available for rootless users.
 
 ## OPTIONS
-**--quiet**, **-q**
-
-The `quiet` option will restrict the output to only the network names.
-
 **--format**, **-f**
 
 Pretty-print networks to JSON or using a Go template.
