+++ pkg/domain/infra/abi/system_novalink.go
@@ -1,0 +1,14 @@
+// +build !varlink
+
+package abi
+
+import (
+	"context"
+
+	"github.com/containers/libpod/pkg/domain/entities"
+	"github.com/pkg/errors"
+)
+
+func (ic *ContainerEngine) VarlinkService(_ context.Context, opts entities.ServiceOptions) error {
+	return errors.Errorf("varlink is not supported")
+}
